refactor: rename/replace pin-input with modular new pin-input
