+++ src/pages/extra-components/index.tsx
@@ -1,14 +1,13 @@
 import { Breadcrumb, BreadcrumbItem } from '@/components/custom/breadcrumb'
 import { Layout, LayoutBody, LayoutHeader } from '@/components/custom/layout'
-import { PinInput2, PinInputField } from '@/components/custom/pin-input-2'
-import { PinInput } from '@/components/custom/pin-input'
+import { PinInput, PinInputField } from '@/components/custom/pin-input'
+import { PinInputOg } from '@/components/custom/pin-input-original'
 import { Search } from '@/components/search'
 import ThemeSwitch from '@/components/theme-switch'
 import { Separator } from '@/components/ui/separator'
 import { UserNav } from '@/components/user-nav'
 import { IconChevronRight } from '@tabler/icons-react'
 import { Link } from 'react-router-dom'
-import { Input } from '@/components/ui/input'
 
 export default function ExtraComponents() {
   const items = [
@@ -52,24 +51,18 @@ export default function ExtraComponents() {
         <Breadcrumb>{items}</Breadcrumb>
         {/* <Input value={pinVal} onChange={(e) => setPinVal(e.target.value)} /> */}
         <h1>Fucking pin</h1>
-        <PinInput
+        <PinInputOg
           autoFocus
           defaultValue='5679'
-          // value={pinVal}
-          // onChange={setPinVal}
           type='numeric'
           length={4}
-          // onComplete={(f) => console.log('complted', f)}
           ariaLabel='Pin Input'
         />
         {/* <PinInput defaultValue='584930' /> */}
-        <PinInput2
+        <PinInput
           className='flex h-10 space-x-4'
           defaultValue='56'
-          // value={pinVal}
-          // onChange={setPinVal}
           onComplete={(str) => console.log('completed', str)}
-          inputComponent={Input}
           autoFocus
         >
           <PinInputField />
@@ -77,18 +70,7 @@ export default function ExtraComponents() {
           <Separator orientation='vertical' />
           <PinInputField />
           <PinInputField />
-          {/* <PinInputField
-            component={Input}
-            className='bg-yellow-50 dark:bg-yellow-950'
-          /> */}
-          {/* <Separator orientation='vertical' />
-          <PinInputField className='bg-orange-50 dark:bg-orange-950' />
-          <Separator orientation='vertical' />
-          <h1>HEllo</h1>
-          <PinInputField className='bg-lime-50 dark:bg-lime-950' />
-          <Separator orientation='vertical' />
-          <PinInputField className='bg-cyan-50 dark:bg-cyan-950' /> */}
-        </PinInput2>
+        </PinInput>
         <div className='flex h-5 items-center space-x-4 text-sm'>
           <div>Blog</div>
           <Separator orientation='vertical' />
